% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foRmulaTV.R
\name{formulatv}
\alias{formulatv}
\title{Scraps formulatv.com and downloads spanish TV audience data}
\usage{
formulatv(ruta, finicio, ffinal)
}
\arguments{
\item{ruta}{The path where you want to create the csv file}

\item{finicio}{The date from which you want to scrap}

\item{ffinal}{The date until you want to scrap}
}
\value{
A data frame will be imported to your workspace and a csv file will be generated in the specified path
}
\description{
Scraps formulatv.com and downloads spanish TV audience data
}
